PMD fixes

+++ src/com/dmdirc/addons/ui_swing/components/durationeditor/DurationDisplay.java
@@ -99,7 +99,7 @@ public class DurationDisplay extends JPanel implements ActionListener,
         super();
         
         this.window = window;
-        this.duration = Long.valueOf(duration / 1000).intValue();
+        this.duration = (int) (duration / 1000);
         listeners = new ListenerList();
 
         initComponents();
@@ -108,7 +108,7 @@ public class DurationDisplay extends JPanel implements ActionListener,
     }
 
     /**
-     * Initliases and lays out the components.
+     * Initialises and lays out the components.
      */
     private void initComponents() {
         button = new JButton("Edit");
